requirements.txt
attrs==17.4.0
certifi==2018.1.18
chardet==3.0.4
click==6.7
decorator==4.2.1
dj-database-url==0.4.2
-e git://github.com/hersonls/djamin.git@5b5887c84341065844ab44b6cfc8ccb865ccfdb9#egg=djamin
Django==1.11.2
django-model-utils==3.1.1
funcsigs==1.0.2
future==0.16.0
geocoder==1.36.0
gunicorn==19.7.1
idna==2.6
Pillow==5.0.0
pluggy==0.6.0
psycopg2==2.7.4
py==1.5.2
pytest==3.4.0
pytest-django==3.1.2
pytz==2018.3
ratelim==0.1.6
requests==2.18.4
six==1.11.0
ua-parser==0.7.3
urllib3==1.22
user-agents==1.1.0
whitenoise==3.3.1
